chore(refactoring): Move export workflow into separate package.
